Fjerne ubrukt kode og fiks renaming conventions
