+++ app/seksjon/din-situasjon/v1/DinSituasjonViewV1.tsx
@@ -1,18 +1,16 @@
 import { ArrowLeftIcon, ArrowRightIcon } from "@navikt/aksel-icons";
 import { Alert, Button, HStack, VStack } from "@navikt/ds-react";
 import { useForm } from "@rvf/react-router";
-import { Form, useActionData, useLoaderData, useNavigate } from "react-router";
+import { Form, useActionData, useLoaderData, useNavigate, useNavigation } from "react-router";
 import { Spørsmål } from "~/components/spørsmål/Spørsmål";
 import { useNullstillSkjulteFelter } from "~/hooks/useNullstillSkjulteFelter";
-import { action, loader } from "~/routes/$soknadId.bostedsland";
-import { bostedslandSchema } from "~/seksjon/bostedsland/v1/bostedsland.schema";
-import {
-  bostedslandSpørsmål,
-  BostedslandSvar,
-} from "~/seksjon/bostedsland/v1/bostedsland.spørsmål";
+import { action, loader } from "~/routes/$soknadId.din-situasjon";
+import { dinSituasjonSchema } from "~/seksjon/din-situasjon/v1/din-situasjon.schema";
+import { dinSituasjonSpørsmål, DinSituasjonSvar } from "./din-situasjon.spørsmål";
 import { useEffect } from "react";
 
-export function BostedslandView_V1() {
+export function DinSituasjonViewV1() {
+  const { state } = useNavigation();
   const loaderData = useLoaderData<typeof loader>();
   const actionData = useActionData<typeof action>();
   const navigate = useNavigate();
@@ -20,7 +18,7 @@ export function BostedslandView_V1() {
   const form = useForm({
     method: "PUT",
     submitSource: "state",
-    schema: bostedslandSchema,
+    schema: dinSituasjonSchema,
     validationBehaviorConfig: {
       initial: "onBlur",
       whenTouched: "onBlur",
@@ -29,25 +27,27 @@ export function BostedslandView_V1() {
     defaultValues: { ...loaderData.skjema, versjon: loaderData.versjon },
   });
 
-  useNullstillSkjulteFelter<BostedslandSvar>(form, bostedslandSpørsmål);
+  useNullstillSkjulteFelter<DinSituasjonSvar>(form, dinSituasjonSpørsmål);
 
   return (
     <div className="innhold">
-      <h2>Bostedsland</h2>
+      <h2>Din situasjon</h2>
       <VStack gap="20">
         <VStack gap="6">
           <Form {...form.getFormProps()}>
             <input type="hidden" name="versjon" value={loaderData.versjon} />
             <VStack gap="8">
-              {bostedslandSpørsmål.map((spørsmål) => {
+              {dinSituasjonSpørsmål.map((spørsmål) => {
+                // Skip rendering if the question should not be shown based on current answers
                 if (spørsmål.visHvis && !spørsmål.visHvis(form.value())) {
                   return null;
                 }
+
                 return (
                   <Spørsmål
                     key={spørsmål.id}
                     spørsmål={spørsmål}
-                    formScope={form.scope(spørsmål.id as keyof BostedslandSvar)}
+                    formScope={form.scope(spørsmål.id as keyof DinSituasjonSvar)}
                   />
                 );
               })}
@@ -58,7 +58,6 @@ export function BostedslandView_V1() {
                 </Alert>
               )}
             </VStack>
-
             <HStack gap="4" className="mt-8">
               <Button
                 variant="secondary"
@@ -72,6 +71,7 @@ export function BostedslandView_V1() {
                 type="submit"
                 iconPosition="right"
                 icon={<ArrowRightIcon />}
+                loading={state === "submitting" || state === "loading"}
               >
                 Neste steg
               </Button>
