--- app/utils/versjon.utils.ts
@@ -1,5 +1,0 @@
-export type SeksjonDataType = { versjon: number; skjema: string };
-
-export function defaultVersjonSvar() {
-  return "Noe gikk galt";
-}
